Fixed issue with having no cancel function for vesting withdrawal and started adding next withdraw date to UI.
