+++ index.html
@@ -1,17 +1,25 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
     <head>
-        <meta charset="UTF-8" />
-        <title>Hello React!</title>
+        <meta charset="UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <title>MUSE</title>
+        <!--<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">-->
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+        <link rel="stylesheet" href="assets/css/custom.css">
+        <link rel="stylesheet" href="assets/css/popup.css">
+
+        <link rel="shortcut icon" href="assets/img/favicon.png"/>
+        <script src="steem-js/dist/muse.min.js">
+
+        </script>
     </head>
+
     <body>
-        <div id="example"></div>
-
-        <!-- Dependencies -->
-        <script src="./node_modules/react/dist/react.js"></script>
-        <script src="./node_modules/react-dom/dist/react-dom.js"></script>
-
-        <!-- Main -->
-        <script src="./dist/bundle.js"></script>
+        <div id="app"></div>
+        <div id="popup_container"></div>
+        <script src="index.js"></script>
     </body>
 </html>
