--- indexMAIN.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>MUSE</title>
-        <!--<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">-->
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css">
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-        <link rel="stylesheet" href="assets/css/custom.css">
-        <link rel="stylesheet" href="assets/css/popup.css">
-
-        <link rel="shortcut icon" href="assets/img/favicon.png"/>
-        <script src="steem-js/dist/muse.min.js">
-        
-        </script>
-    </head>
-
-    <body>
-        <div id="app"></div>
-        <div id="popup_container"></div>
-        <script src="index.js"></script>
-    </body>
-</html>
