+++ pages/Wallet.jsx
@@ -53,7 +53,7 @@ class Wallet extends Component {
         this.transferFct = this.transferFct.bind(this);
         this.transferToVestFct = this.transferToVestFct.bind(this);
         this.withdrawVestingFct = this.withdrawVestingFct.bind(this);
-        
+
         this.toggleModalTransfer = this.toggleModalTransfer.bind(this);
         this.toggleModalTransferVesting = this.toggleModalTransferVesting.bind(this);
         this.toggleModalWithdrawVesting = this.toggleModalWithdrawVesting.bind(this);
@@ -183,7 +183,7 @@ class Wallet extends Component {
                 {
                     alert(tmpSplit[0].split(".")[0]);
                     return true;
-                }           
+                }
             } else {
                 alert(lastErrorMsg[0]);
                 return true;
@@ -194,12 +194,12 @@ class Wallet extends Component {
             return false;
         }
     }
-    
+
     callbackTransferFunds(code, message) {
         if(code == -1){
             if(!this.tryHandleError())
             {
-                alert('Unable to transfer funds');    
+                alert('Unable to transfer funds');
             }
         } else if(code == 1){
             // success
@@ -219,7 +219,7 @@ class Wallet extends Component {
         if(code == -1){
             if(!this.tryHandleError())
             {
-                alert('Unable to transfer funds to vesting');    
+                alert('Unable to transfer funds to vesting');
             }
         } else if(code == 1){
             // success
@@ -303,7 +303,7 @@ class Wallet extends Component {
 
     transferToAccountProc() {
         let isValidated = true;
-        
+
         if(this.state.toAccount == ""){
             isValidated = false;
             alert("Please input the account name ");
@@ -317,7 +317,7 @@ class Wallet extends Component {
             alert("Please input the sufficient funds in the amount field.");
         }
 
-        
+
         if(isValidated){
             this.setState({loading: true});
 
@@ -340,7 +340,7 @@ class Wallet extends Component {
             alert("Please input the sufficient funds in the amount field.");
         }
 
-        
+
         if(isValidated){
             this.setState({loading: true});
 
@@ -362,7 +362,7 @@ class Wallet extends Component {
             isValidated = false;
             alert("Please input the sufficient vesting in the amount field.");
         }
-        
+
         if(isValidated){
             this.setState({loading: true});
 
@@ -371,6 +371,14 @@ class Wallet extends Component {
             muse.withdrawVesting(username, pwd, Number(this.state.amount).toFixed(6), this.callbackWithdrawVesting);
         }
         //this.toggleModalTransferVesting();
+    }
+
+    withdrawVestingCancelProc() {
+
+    let username = window.localStorage.getItem('username');
+    let pwd = window.localStorage.getItem('password');
+    muse.withdrawVesting(username, pwd, Number(0).toFixed(6), this.callbackWithdrawVesting);
+
     }
 
     defaultHistoryFormatter(userName, operationName, date, operationData, additionnal) {
@@ -424,7 +432,7 @@ class Wallet extends Component {
             history_info.text = 'Withdrawing ' + operationData.vesting_shares.split(" ")[0] + ' VEST';
             break;
         case "account_witness_vote":
-        
+
             if(operationData.approve)
             {
             history_info.text =  operationData.account +' Voted Witness ' + operationData.witness;
@@ -449,7 +457,7 @@ class Wallet extends Component {
         case "custom_json":
             history_info.text = 'Custom Json ' + operationData.id + ' ' + operationData.json + ' ' + operationData.required_auths + ' ' + operationData.required_basic_auths;
             break;
-        default: 
+        default:
             history_info.text = 'Unknown operation: ' + operationName;
         }
 
@@ -485,15 +493,15 @@ class Wallet extends Component {
     }
 
     componentWillReceiveProps(nextProps) {
-        
+
     }
 
     componentDidMount(){
-        
+
     }
 
     componentWillUnmount(){
-        
+
     }
 
     componentDidUpdate(){
@@ -518,11 +526,11 @@ class Wallet extends Component {
                             <div className="col-md-9">
                                 <h3>MUSE</h3>
                                 <div className="text-muted">
-                                    Tradeable tokens that may be transferred anywhere at anytime.<br/>MUSE can be vested by clicking transfer to Vesting. 
+                                    Tradeable tokens that may be transferred anywhere at anytime.<br/>MUSE can be vested by clicking transfer to Vesting.
                                 </div>
                             </div>
                             <div className="col-md-3 text-right margin-top-30">
-                                
+
                                 <Dropdown isOpen={this.state.dropdownOpen} toggle={this.toggle}>
                                     <DropdownToggle caret>
                                         {this.state.walletInfo.balance} MUSE&nbsp;&nbsp;&nbsp;
@@ -544,14 +552,16 @@ class Wallet extends Component {
                                 </div>
                             </div>
                             <div className="col-md-3 text-right">
-                                
+
                                 <Dropdown isOpen={this.state.dropdownOpenVM} toggle={this.toggleVM}>
                                     <DropdownToggle caret>
                                         {this.state.walletInfo.vesting} MUSE&nbsp;&nbsp;&nbsp;
                                     </DropdownToggle>
                                     <DropdownMenu>
                                         <DropdownItem onClick={this.withdrawVestingFct}>Withdraw from Vesting</DropdownItem>
+                                        <DropdownItem onClick={this.withdrawVestingCancelProc}>Cancel Withdraw from Vesting</DropdownItem>
                                     </DropdownMenu>
+
                                 </Dropdown>
                             </div>
                         </div>
@@ -575,6 +585,15 @@ class Wallet extends Component {
                             </div>
                         </div>
 
+                        <div className="row form-group margin-top-30">
+                            <div className="col-md-9">
+                                <h3>Next Withdrawal</h3>
+                                <div className="text-muted">
+                                    
+                                </div>
+                            </div>
+                        </div>
+
                         <div className="margin-top-50">
                             <h3>HISTORY</h3>
                             <table className="table table-hover table-striped">
@@ -592,7 +611,7 @@ class Wallet extends Component {
                                 </tbody>
                             </table>
                         </div>
-                        
+
                     </TabPanel>
                     <TabPanel>
                         <div className="row form-group margin-top-30">
@@ -675,7 +694,7 @@ class Wallet extends Component {
                         </div><hr/>
 
                         <div className="margin-top-50">
-                            
+
                             <div className="form-group margin-top-50">
                                 <label for="account_name">ACCOUNT NAME / MuseID</label>
                                 <input type="text" className="form-control" id="account_name" value={this.state.walletInfo.userName} disabled />
@@ -718,13 +737,13 @@ class Wallet extends Component {
 
                             <div className="form-group margin-top-50">
                                 <button type="button" className="btn btn-primary" onClick={this.updatePassword}>UPDATE PASSWORD</button>
-                            </div>                            
-                        
+                            </div>
+
                         </div>
                     </TabPanel>
 
                 </Tabs>
-                
+
                 <Modal isOpen={this.state.modalTransfer} toggle={this.toggleModalTransfer} className={this.props.className}>
                     <ModalHeader toggle={this.toggleModalTransfer}>Transfer to Account</ModalHeader>
                     <ModalBody>
@@ -759,7 +778,7 @@ class Wallet extends Component {
                                 <input className="form-control" type="text" id="transfer-memo" onChange={this.changeMemo} />
                             </div>
                         </div>
-                        
+
                     </ModalBody>
                     <ModalFooter>
                         <Button color="primary" onClick={this.transferToAccountProc}>SUBMIT</Button>{' '}
