+++ webpack.config.js
@@ -1,34 +1,53 @@
-module.exports = {
-    entry: "./src/index.tsx",
-    output: {
-        filename: "bundle.js",
-        path: __dirname + "/dist"
+var webpack = require('webpack');
+var path = require("path");
+
+var config = {
+    entry: {
+        app: ['./main.js']
     },
 
-    // Enable sourcemaps for debugging webpack's output.
-    devtool: "source-map",
+    output: {
+        path: path.resolve(__dirname, "/"),
+        filename: 'index.js',
+    },
 
-    resolve: {
-        // Add '.ts' and '.tsx' as resolvable extensions.
-        extensions: [".ts", ".tsx", ".js", ".json"]
+    devServer: {
+        historyApiFallback: true,
+        inline: true,
+        port: 8080
     },
 
     module: {
-        rules: [
-            // All files with a '.ts' or '.tsx' extension will be handled by 'awesome-typescript-loader'.
-            { test: /\.tsx?$/, loader: "awesome-typescript-loader" },
+        loaders: [
+            {
+                test: /\.jsx?$/,
+                exclude: /node_modules/,
+                loader: 'babel-loader',
 
-            // All output '.js' files will have any sourcemaps re-processed by 'source-map-loader'.
-            { enforce: "pre", test: /\.js$/, loader: "source-map-loader" }
-        ]
+                query: {
+                    presets: ['es2015', 'react'],
+                    plugins: ['transform-object-assign']
+                }
+            }
+        ],
+        exprContextCritical: false
     },
 
-    // When importing a module whose path matches one of the following, just
-    // assume a corresponding global variable exists and use that instead.
-    // This is important because it allows us to avoid bundling all of our
-    // dependencies, which allows browsers to cache those libraries between builds.
-    externals: {
-        "react": "React",
-        "react-dom": "ReactDOM"
+    node: {
+        fs: "empty",
+        net: 'empty',
+        tls: 'empty',
+        dns: 'empty'
     },
-};
+
+    plugins: [
+        new webpack.DefinePlugin({
+            'process.env': {
+                //'NODE_ENV': JSON.stringify('development')
+                'NODE_ENV': JSON.stringify('production')
+            }
+        })
+    ]
+}
+
+module.exports = config;
